Fix subset of global virtual grids (#6143)
When virtual grids are passed and we need a subset, we were stymied by the first test for the rotation of a global grid.  In reality, we may need to do both rotation and cropping but the first test was chained with an else-if and hence the subset was skipped.  Closes #6140.