Remember separate fill settings for lakes and river-lakes
See $6038 for background.  The problem was that we considerer Ctrl->C.active when setting the lake and river fill but in this case only the river fill was specified and the lake should default to -S, which here is no fill.
Improved readability of code by introducing named enums instead or integers 0-5 in the fill array.