Catch a SEGV due to bad file name
I got a nasty SEGV because I mistyped -W1p as =W1p since we made the assumption that = was a grid format (e.g., test.grd=bf).  But here we simply chopped lff the entire file and a file of NULL caused the SEGV.  This PR detects this case and undoes the chop since we just don't know what the file is and if no such file we get the benign message
psxy [ERROR]: Cannot find file =W1p
instead.