Make checking for registrations more robust (#6160)
THe gmt_set_unspecified_remote_registration function, whose job it is to determine of the pixel or gridline registration was specifically requested in a remote file name, was written before there was @earth_geoid and hence the lame check for _g or _p found _g(eoid) and returned prematurely.
This PR adds the harder work of advancing past the leading dataset name and then look for _g or _p.
In the process I learned I had a typo in the test gmt_data_server.txt where the earth_relief_01d_g.grd had a missing / in front of server.  Just delete your test/server/gmt_data_server.txt to get the fixed file.

Tests pass.