Address a series of issues related to vectors (#6161)
There have been several issues identified in GMT vectors lately (e.g., #6108) and we have had some partial fixes.  This PR reconciles several problems and completes started enhancements:

The new minimum vector shrink in +mnorm/limit was stated to be 0.25 but in fact was never set so it was still zero.  This is now fixed.
The check for geovector to skip heads if exceeding the vector length was flawed so some small vectors ended up with bad heads (e.g., #6108).  Now the test is fixed and heads exceeding size is not plotted.
Cartesian vectors never bothered to check and hence always plotted the heads even though exceeding vector length.  This has now been addressed.
FInally, there are uses of vectors where the vector length should not be considered when plotting heads. Many of these are internal (e.g., adding a vector head at the end of the line or at the end of a math x-axis), but placing flow vectors on a graph may also fall prey to the limitation.  Using +n without any argument means always plot the head regardless of vector length.

These corrections meant that 8-10 PostScript plots needed to be updated. For some I chose to use +n to avoid having to change the  plot, while others would benefit from a refresh (e.g., properly giving scaling in km for geographic bectors in sample.sh).

Closes #6158 and #6108.