Ensure small circle vector lengths can use map units
See #6166 for background.  This PR fixes mistakes in both plot and plot3 where the case for small circle and length was not considered.
Instead of adding a new test I simpy added k to the values in smallcirclevec.sh and changed one to be in meters.  All tests work as before plus the problem was fixed.  Closes #6166.