Check if plotting degenerate symbols for legend
When plotting circles on a map with -SE- we are plotting circles but doing so via the elliptical outline option to get a projected circle.  In those cases, if we use -l we do not want to see an ellipse in the legend but a circle.  Same for -SJ- for degenerate rectangles as squares.