Switch to using q instead of u for user unit
This change is driven by the fact u is already one of the recognized map distance units (survey foot) and hence we want a unique letter to indicate 'user unit'.  Such units are used in a few places, for instance the width of a vertical bar may be specified in x-units and for columns it may separately be specified as x and/or y-units.

Fixed missing usage message for teh 3-D cube symbol in psxyz and made sure all tests worked with the old 'u' before changing that to 'q' and rechecking the tests.