Let grdedit -D+c allow setting/removing default CPT for grid (#6223)
Given the cpt is now one of the attributes, we need a way to add, remove, or change it via grdedit.  This PR adds +c[-|cpt] to the -D option.  If no argument is used then we add the default CPT (turbo), if argument is - we remove any default CPT from the grid, else we add the specified CPT.  The PR also fixed a bug that failed to consider NaNs as the NaN values and thus lead to uncessisary reading.