Fix regression on parsing MON-dd-yyyy strings
Due to a regression some time ago, we would return an error in gmt_scanf when a string did not start with a number.  This would catch valid dates like Jun-01-2014 and get thrown out before we even get to parse it given the correct FORMAT_DATE_IN.  This PR only allows the touch check if not known to be ABSTIME and then we can at least fail in the GMT_IS_ABSTIME branch.
Also removed the unused no_T variable set if we find a date with missing trailing T, since we did not do anything about it (and parsing of the date allows it).