Improve openMP variables and macros to avoid warnings
THe gmt_M_grd_loop and similar macros result in tons of warnings due to the problem with Windows not supporting OpenMP > 2.0 and forces us to have signed loop variables.  This PR addresses this by introducing a typedef openmp_int that is set accordingly, and we use casts to keep things quiet regardless of what openmp_int is.