Improved handling when remote data are unavailble (#6231)
In testing the updated remote file I made a mistake in moving the comment with the GMT version to another line.  Unfortuntely, it was expected to be in teh 2nd line.  THe resulting error handling was insufficient to prevent a crash.  This PR ensures that if we have a failure to parse the data or index files then we gratiously report and error and prevent further access to these files during this process.