Ensure consistent use of table, t, +t when discussing tables
See #6236 for background.  This PR ensures we use directives t or modifier +t when referring to tables instead of f and +f for files in some places.  Because of the mixup we will honor f and +r under the hood for backwards compatibility but have updated the usage message and docs.  Closes #6236.