Read dcw.conf first so it takes precedence over dcw-collections.txt (#6244)
We want the user to have the ability to add their own collections and even redefine what any particular tag also present in the sytsem file should mean.  If I want Europe to include the Bouvet Island in the S Atlantic then I am allowed to give a crazy -R setting for that name.  This requires we read the user dcw.conf first since the search is linear.