Let coupe parse old format without error
See the forum post for backtround.  Basically, older GMT versions did not bother to check if the sscanf was successful, and when encountering 50k instead of 50 it gave up.  In GMT 6.2 we added more checking so this resulted in a parsing error but no message was given.

This PR now gives a warning if scanning is unsuccessful but it does not exit the program.