Avoid race conditions when we can
The mktemp is generally not recommended since it only creates a unique filename template, but by the time we use it to create a file it may no longer be unique (another user could have created that file, somehow).  However, when creating a directory there is mkdtemp which sets the template and opens the file at the same time, avoiding this race condition.  Unfortunately, not available on Windows so keeping the old scheme for WIN32.

The final place we use mktemp to create a unique file name is in grdblend and here we do not want to open the file since it is an output gridfile name passed to a GMT module.  We will simply have to tolerate intolerant rants from the compilers.