Speed up grdmask -S (#6253)
* Speed up grdmask -S

When we use grdmask -S we basically search all nodes within the specified radius of a data point and set then to inside.  However, for denser data and/or larger radii, this results in the same nodes being visited multiple times and each time we need to compute a radial distance to see if we are inside or not.  This PR simply adds a char array to keep track if we have flagged a node as inside and if we visit it again we immediately skip to the next node.

* Update grdmask.c