Exclude grdimage padding when sett CPT limits
See #6218 for background.  This had its origin in trying to limit the extra rows/columns used in order to allow spline resampling near the boundaries.  We had added a check that turned bicubic sampling off to avoid extending the region and getting z-values that could bias the min/max range and hence the CPT stretching.  This PR removes that check (we do want bicubic sampling as default) but recomputes the grid zmin/zmax to only apply to the regions specified on the command line.  Lots of "failures" because we are undoing the previous decision but this is a better fx.