Allow -S to take appended map unit
For the units that accept dimensions in map units (geo ellipse and rotated rectangle), users can indicate the unit of the data (if read from file) rather than having to append the unit to the columns in the data.  E.g., -SEd means that the major and minor axes read from file will be in spherical degrees and not km.