Let the debug message finding -R -J history report the information
In modern mode we go hunting for -R -J values in the history if -R -J are not specified on the command line.  If we find these in the history we append -R and/or -J to the internal command and report the lame message:

Modern mode: Added -R to options since history is available

However, the intended target for this message (developers) would appreciate to actually be told what was found, e.g.

Modern mode: Added -R to options since history is available [0/40/-2/5]

This PR takes care of that.