gmt inset shall not accept -X -Y
While the -X and -Y options serve no purpose in the gmt inset call, and are not listed as allowable options neither in synopsis nor documentation, we did in fact not prohibit them, leading to troubles (#6288).
This PR slams that door shot.  No example or tests used -X -Y since that was never the plan.