Fix limitation in -J parsing in inset and subplots
In these situations, the user may now know the actual dimension of the plot area, hence it is awkward to specify a map width or scale.  THe standard way is to leave off the scale or width entirely (internally, we then append ?).  However, two projections (-Jo and -Jg) have grown a series of optional modifiers and these were not considered when looking for the end of teh main projection parameters.  This PR checks these cases, strips off the modifiers, does the check and possible appenidng ?, then adds any modifiers found.  Hence, expressions like -JG30/15/+w45 is now correctly parsed to be come -JG30/15/?+w45 and all is well again.