Must let inset process -X -Y (#6304)
Because there are things happening under the hood that adds -X -Y to a gmt inset call, we need to keep XY among the list in THIS_MODULE_OPTIONS, thus reverting #6298.  I will see if I can find a way to distinguish between these GMT-imposed -X/-Y settings and that of a user on the command line (which we would like to prevent).  For now, just reverting to avoid errors.