Let GMT verify counters in external datasets
As a sanity check, we let GMT validate the counters found in a dataset passed to a module vi reference or duplicate.  Mismatches between stated and actual records and segments and columns will be noted and an error is returned.
In the processes I found pslegend did not initialize n_segments to zero in the dataset and thus got caught.