Clarify how -Z in a data header affects color
This pertains to issue #6312 and lines or polygons.  With -G it is not an issue - a filled polygon is secured.  However, for -Z we do not know if it is a color change for a polygon or a line.  By default we use the color for the line.  This can be further changed via -L (fill polygon instead) or -W+c (set the color for fill and line).