Let default remote grid registration be gridline for grdtrack (#6319)
See #6299 for background.  This PR ensures that grdtrack first selects a gridline-registered version of a remote grid unless registration is specified by the user.  This ensures that sampling of arbitrary points is optimally handled.  The change is documented in the man page.  Also, since we now no longer adds a grid pad for non-bicubic sampling we must allow for a node outside the range when the input point lands on a node.  This is not an error (since the weight there is zero) so we simply skip rather than cause a crash via assert.  Closes #6299.