Let grdblend accept inner blend region with or without leading -R (#6332)
It is easy to forget that the blend file region needs a leading -R before the w/e/s/n, and since I did forgot I ended up with the wrong results (no inner border and no tapering).  This PR makes that -R optional and strengthens the checks involved.  It also fixes a potential bug in that the variable n (number of grids read so far) was confused with n_scanned in line 268 (now fixed).
I will update the documentation for this later as I am working on a figure to help blend users.