Allow for tiny slop when computing pads in grdimage
Due to the need to add extra rows/cols for some projections and interpolants, we read in extra data from a file if it has it.  Hwoever, when an automatic CPT scaling is needed we do not want to include this extra pad data in determining the zmin/zmax limits so we temporarily scale back and recompute the min/max in those cases.  However, the calculation that determined the pad was subject to round-off and did not follow what is done elsewhere in GMT (tolerate 10e-12 in such cases.  So 1.9999999 became 1 instead of 2, for instance.

This PR fixes this issue.