Check for angles that are either 0, -180 or 180 in gmtplot_get_outside_point_extension (#6371)
This function is responsible for extending a line that intersects the rectanbular box so that the width of the line is taking into consideration (if we just stop drawing exactly at the intersection coordinates then the finite width of an oblique line will leave a gap).  The case highlighted in this bug was when the line was identified to cross the left boundary (the line goes through the top-left corner) but it could also be seen as crossing the top boundary (at 90 degrees).  Either case should result in no extension, but since it was seen to cross the left bouneary (when in fact the line is parallel to it) we did not check for that case and ended up with a crazy value.