Fix the lack of precision in -Ib
The -Ib outputs the bounding box of the grid.  But it failed to do so for the increment resolution and instead only output the 5 corner points,  This mean that projecting this polygon would failed to outline the polygon in an oblique projection.,

This PR samples the perimeter at the grid spacing.