Let grdsaemount -Co handle u = -1
For u = -1 in the polynomial shape we end up with a 0/0 expression which evaluates to NaN.  However, we know from l'Hopital's rule that the result should beo 0.  This PR intervenes for -1 to yield 0.  WHile at it, since this function and its gradietn are only defined over ±1 I check r and set results to 0 if outside.