Return with error if libcurl fails (#6479)
If we request a remote file that does not exist we would get a long list of repeated error messages about failures and suggestions to turn off internet download, etc.  Part of the problem was the error code was not returned, so we tried again later on in the module.  Second, the message

grdimage [WARNING]: You can turn remote file download off by setting GMT_DATA_UPDATE_INTERVAL to "off"

is really not that helpful in this situation (it is better if there is time-outs), and third, the general HTTP error says nothing, so it is unclear what is happening.

This PR makes these changes:

1. IT properly returns the error code so the module can gracefully stop
2. It no longer prints that misplaced message about GMT_DATA_UPDATE_INTERVAL
3. It suggests that if this error is returned that it likely means the file cannot be found.