Fix use of data offset when grdpaste reads grids (#6488)
Sometime during the recoding from GMT4 to GMT 5 we introduced a bug that only affected grdpaste.  THe way grdpaste works is to set a data offset parameter in the hidden grid header section so that when we read in data we consider if the hidden data_offset is nonzero (well, we basically always offset the start of the grid by that amount (zero or not).  Thus data are read by grdpaste nto the same array but at two different starting positions.  This works fine.
The trouble had to do with input grids that has a scale factor and offset that are not the default (1, 0).  Turned out the code that does the unscaling was not passed the data offset and hence only the first part of the combined grid would have scaling applied.
This PR applies the offset immedieatly at the top of gmtlib_read_grd and then all subsequent function calls do not need to consider any offsets.