Let -N level default to minimum value
Rather than have to look up what that might be for a grid, just giving -N with no value specified will no longer give the error (missing argument) but set the level to the minimum value in the relief grid.  If -R was also set then we pick the lowest value of -R and the grid.