Ensure auto-range in histogram covers the bins
See forum post here for background.  The trouble was that depending on tick interval (or default interval) we may not round min/max to include the last bin or the two centered bins if -F.  This PR ensures the raw min/max determined from the data is extended by the actual bin width (and dealing with -F) so that the further rounding due to tick intervals will always at least show all the bins.