Improve histogram handling of time units via -T (#6509)
* Improve handling of time units via -T

The -T array option allows specification of time units (e.g., 1o, 7d).  Internally, system default units are controlled by TIME_UNIT but it is pretty awkward (and hard to remember) to also set TIME_UNIT to the same as the -T unit.
This PR adds a check during the parsing of -T to detect the presence of a time unit and then re-scales the min/max values found from whatever initial units GMT used (typically seconds) to the chosen unit.  If -R has already been parsed it also scales those values to the new units which are now used internally for this process.

* Add new function for the re-scaling

* Update explain_array.rst_

* Update explain_array.rst_