Add revised 2014 x2sys paper via ResearchGate URL (#6515)
The x2sys software got an update in 2014 after the major revision in 2010 turned out to have a few weaknesses.  The accepted final revision Word document from 2019 was updated with the corrections and posted on ResearchGate and I have added links to this site from the documentation.  This makes it easier for people who do not have access to the journal to follow what is going on in x2sys.

Closes #6412.