Allow contour files to only list contour level
Since it is awkward to create a list of desired contours and add the character C to the end, this PR allows a plain single-column file of contour levels to be accepted in -C by grdcontour and pscontour and the type defaults to plain contour (C) only. Currently we just get an ugly read error.