Improve clip path for images in -Qi (#6523)
* Improve clip path for images in -Qi

THe algorithm used for building the staircase clip path for the image in grdview -Qi was somewhat smart but not smart enough, leaving a path when there is no valid image to the far left or right.  WHile clipping is fine, the fact that the path extends too far means ghostscript senses these and the bounding box is too large.

This PR eliminates that problem.  I also added more comments to explain the otherwise cryptic section.  THe plot below show an actual 3-D image clip path.  The heavy line was the previous version and the new version is the thin yellow line.

No tests were affected.

* Update grdview.c

* Update grdview.c