Address bug in NORM operator
The NORM operator in g??math says it will convert A so that max(A)-min(A) = 1.  This it does, but not it a good way.  For instance, if your data set generally goes from 3.0 to 3.5 and you call NORM, you get a result that goes from 6-7 instead - hardly a useful transformation.  What is needed of course is to say that NORM converts A so that min(A) = 0 and max(A) = 1.  This bug (and bad documentation) has now been fixed for both grdmath and gmtmath with this PR.