Fix -R parsing check for geographic data
We correctlyl detect if the -R string is using recocognized longitude and latitude formats, but then we had a bad check were we needed a bit-wise oeprator obut instead did an equal, i.e.

col_type == GMT_IS_GEO

instead of

col_type & GMT_IS_GEO

and we did this because we called a macro to get the type and hence did not try the bitwise and I guess.
This now sets geographic internally.