Fix regression in parsing of -W
See #6538 for the problem.  We had a regression in all the blockm* modules where a -W+s argument would yield a parsing error but not tell anyone - just exiting.  Unfortunately there is no workaround since adding i or o turns off the opposite direction; here the user wants weights in and sigma out.
Closes #6538.