Fix documentation and parsing of -W (again) for block modules
Despite the earlier fix () there was still more to do.  This PR does the following:

Standardizes the -W parsing via a common BLK_parse_weight_opt that also looks for +w
Fixes glitches in the blockmean documetation (had 1/w instead of 1/w^2)
Updates the language for -W in the docs
Presents the direectives and modifiers in -W in the usage
LIke mode and median, lets -W+s reverse the internal weight back to sigma on output.