Fix error in setting max half-width in node search
When we need to access all nodes inside a search radius, we first determine limits on the number of rows and columns away from the center.  For geographic data this also means we consider the change in longitudinal distance with latitude.  All this works fine.  However, we had a "safety valve" that set the upper limit of how many columns we may want to search to either side of a node to n_columns/2.  This is an error: If you are at the far left column and have a search radius that equals or exceeds the longitude range then you want all the columns to be inside, hence the upper limit must be n_columns.
No tests affected but I ran into this with a grdseamount  research case where I only used a small region to view a portion of a seamount whose radius exceeded half the width and things got oddly clipped...