Fix usage of gmt_access
This general function returns nonzero if there are problems finding the given file.  However, it already does a few other things:

If file is a memory file it returns 0 since memory is there
If file is a cache file then it will first check if we should download it

Hence, there are no points checking both for access and if cache or memfile in various places since gmt_access handles this.  This PR cleans that up a bit.