Three things on a lazy Thursday (#6553)
* Fix usage of gmt_access

This general function returns nonzero if there are problems finding the given file.  However, it already does a few other things:

If file is a memory file it returns 0 since memory is there
If file is a cache file then it will first check if we should download it

Hence, there are no points checking both for access and if cache or memfile in various places since gmt_access handles this.  This PR cleans that up a bit.

* Please JL on the flushing

* Fix gmtapi_extract_argument for the colon options

Needed a special section to handle the quoted, decorated, or contoured line settings.

* Update gmt_api.c

* Update gmt_api.c

* Update gmt_plot.c

* Update gmt_plot.c