Streamline parsing of required arguments
THe module parsers are full of situations where we must obtained a required argument after the option or a modifier.  We thus have lots of repeated checks to see if there is an argument or not, then parse it.  For file names, we also have preliminary checks for existence of input files.  This goes on in 150 modules.  In many places we do not perform a check if an argument was given before converting the string, and in other places where no argument is expected we do not check if one was erroneously given.

This PR starts a new path forward where we place most of the checking and reporting of errors in small sub-functions that returns an error code or 0, and passes any value obtained via a pointer.  This drastically reduces the number of lines in the parsers and consolidates the checks in just a few places.