Fix incorrect auto-offset for centered roses (#6561)
* Fix incorrect auto-offset for centered roses

See #6213 for background.  The trouble was that the auto-offset added in modern mode should of course be zero if the horizxontal or vertical justification is centered - but it was not.  Then, due to rpemature blindness I committed the wrong pplots to dvc so that these errors passed.

This PR fixes the offset calculation and updates the three PS files that were just wrong.

* Update gmt_plot.c