Improve macro checking if option given more than once
We currently have this check for all options taht can only be set once:

				n_errors += gmt_M_repeated_module_option (API, Ctrl->I.active);
				Ctrl->I.active = true;

This PR combines those statements so that only

				n_errors += gmt_M_repeated_module_option (API, Ctrl->I.active);

wil lbe needed, allowing us to remove a few hundred lines and also be sure that active will be set to true in each case.

This PR updates the macro but does not remote the (now) superflous active = true lines in the modules - that will come next.