Address minor lapses due to parser updates
I few issues was notices, and fixed:

The -F option in mapproejct and grdproject has a default setting of e for meter.  IT is clear from usage and documentation, but the code never set the default hence I got confused.
x2sys_init man page said that -G takes a required d or g, but the code and usage messages disagree and I got confused.  Now we allow -G with no argument and the usual default to g
pssac had checks for valid arguments to E and F deep in the code, and my recent changes causes an invalid memory access.  Now I check for valid argument in the parser instead and handles any bad args.

With those we are back to passing all tests again.