Simplify the checks for only one output file
We had lots of cases where we used a counter to count how many times an output file was specified but we now have the new macros that can do this for us.  So saved many lines on this and eliminated unnecessary n_files variables.
Also fixed a few glitches were I forgot to use the macro.