Let movie set an environmental hint we are building frames (#6652)
As I was making an animation I got annoying messages and wrong decisions being made by GMT: 					GMT_Report (GMT->parent, GMT_MSG_WARNING, "Changing to PostScript landscape orientation based on your plot and paper dimensions, but we cannot be 100%% sure.\n");
					GMT_Report (GMT->parent, GMT_MSG_WARNING, "Use PS_MEDIA and/or PS_PAGE_ORIENTATION to specify correct paper dimensions and/or orientation if our guesses are inadequate.\n");

This should never apply when building a movie since we fix the cancas.  Because movie scripts are independent the solution is to set an environmental parameter (GMT_MOVIE=YES) and check thatn in this ase.