Fix some issues related to geo vs cart (#6646)
This PR started as a fix to the problem reported via #6587, but then we had some trickle-down issues. Basically, users have always had the option of using -f to override what a grid file says it is.  Typically, one may read a Cartesian grid that perhaps should be seen as geographic, and then using -fg makes it so.  However, the opposite turned out not to be true.  If you have a geographic grid but wanted it to be treated as a Cartesian grid then -fc did not do that since -fc is parsed first and then when teh geographic grid is read we recognize it as such and basically -fg is set.
The solution is to remember that we set -fg or -fc and then apply it.  This PR checks at the end of gmtapi_import_grid to see if we need to impose one of them, as well as the start of gmtapi_export_grid to see if a change should apply to the output.  In doing this I learned taht GMT_Get_Values, which is this clever API function taht understands units, temporarily may set -fg but then leave it as is.  So in grdflexure, after parsing -E5k, the k for km triggered -fg and then the 0/600/0/600 region was seen as too big for geographic...
WIth these fixes, all seems OK.