Enhance mapproject to give closed domain polygons (#6673)
* Enhance mapproject to give closed domain polygons

This PR follows up a bit on the one earlier today () by adding an optional modifier +n to -We or -Wr.  This will build a closed polygon outline of the oblique regions.  For -Wr that is the interior oblique region while for -We it is the encompassing oblique region.  I have then added a documentation figure to mapproject -W to explain the purpose of these and why they are different.

* Update GMT_obl_regions.sh

* Close polygon exactlu

* Update mapproject.rst

* Update gmt_api.c

* Remove dumb duplicate

* enforce exact fist = last point

* Add Out file variable, check, and init

Also fix bad mode