Demux output grid from grdfft before returning in memory (#6664)
* Demux output grid from grdfft before returning in memory

Normally, the demux takes place in the writing section but here we simply want to return the result to the calling enviroment as a grid.  THus we must demux the grid first.

* Improve the GMT_FFT function

SInce 99% of applications (and 100% in GMT) wants a final real-valued grid back after the inverse transform, we make that the default treating - i.e., demux, eliminate the no-longer-usable imaginary component and reallocate to fit the resource.  Should some developer not want that then they may add GMT_FFT_NO_DEMUX to the mode flag.  By strengthening the default behavior we are better able to pass results out to calling environments.

* Update gmt_enum_dict.h

* Add GMT_FFT_Reset

* Update api.rst

* Update api.rst

Co-authored-by: Joaquim Luis <jmfluis@gmail.com>