Add splines figure and script (#6656)
* Add files via upload

This PR is related to issue #6643. It add a new figure and script to better explain the splines in the docs.

Here I add the script. Let me know if it is ok. 

In what part of the docs should the figure go? At the end of -F argument?

* Update GMT_splines.sh

Updated script following Paul's suggestions.

* Update GMT_splines.sh

Changed name and format of the output file.

* Set executable permissionm

* Update images.dvc

* Update sample1d.rst

* Update GMT_splines.sh

Fix whiespace

* Update images.dvc

Co-authored-by: Paul Wessel <pwessel@hawaii.edu>