Fix wrong counter used with status array in movie and batch
Because of the +soffset modifier to -T, there are two typoes of running frame/job numbers: Those taht go from 0 up and those that go from offset and up.  I mistakenly used the latter as argument to the status array and hence we would exceed array bounds.  Not noticed until this case on the forum since +s has not been exercised by anyone I guess.