Improve documentation and usage for movie
The various fading durations in movie all have default values but the usage and documentation did not indicate that.  That is now fixed.  Also, I have tried to make the  documentation easier to follow by breaking options with multiple modifiers into a list of modifiers and their actions instead of an endless paratraph.