Improve events documentation
LIke for movie, this PR breaks up the long paragraphs for some options with bullet lists for each directive or modifier.
Also, given name of directives and shorthands I decided to change -Mz to -Mv since we refer to value.  The change is backwards compatible as I will look for both -Mz and -MV for the same thing, as well as +z and +v in the -A option.