Fix wrong pen for contour legends
The auto-legend code for grdcontour and pscontour made an assumption that if only one label was given then it was for the annotated contour.  But if the user only sets -C10 and -Wc0.5p,red we want to see a red pen since -A was not specified.  This PR fixes this problem.  No tests were affected.