Fix regression in triangulate and clean up memory mamagement
Using -G and -N at the same time resulted in no node info being written.  This is the bug that is fixed here.
In the processes, I eliminated massive repeated code on freeing memory following an error.  I replaced this with a goto statement that goes to the end to finish up.