Clarify grdflexure listing and address related issues
This PR follows up on a distant grdseamount issue related to the listing output of -L.  It was not defined.  However, what is required for the research use is to write as a numerical column the time in years, followed by trailing text that has two words: File name and time tag.  This matches the concept in grdseamount where the numerical time must be written as well.  The format is now clearly described in the documentation.

This change broke some weak spots elsewhere that had to be fixed:

grdtrack -G+lfile can read a list of files but it was done via a local fopen and fscanf.  Hence, it is vulnerable to files with leading numerical contents.  This PR improves this to select the first word in the trailing text.
Two test scripts showing firmoviscous flexure calcuations nad loops over the content of these undocumented lists.  Now that the contents is known and has a leading numerical value those loops had to be adjusted.

With those changes all tests work.