Correct input machinery for grdflexure
THe documentation talked about providing a list of input grids via =flist, but that does not work since the list has both time and grids. By the time grdflexure parser gets into the action the GMT_Create_Options has already replaced =flist with what it finds in the list and that gives crazy filenames composed of a record with numbers, tabs, and filenames.

This PR not only clarifies what the input options are (and updates the documentation), it no longer allows the =flist mechanism (since it is inappropriate for this modules) and adds flist+l as one mechanism to pass a list (like grdtrack) but also will allow a flist that ends with extension .lis to be recognized as such.