Improve memory cleanup in grdflexure
As in many other module, if we get an error we have lots of thing to free up, so it is more exonomical do to so at the end via an emergency goto statement.
Also fixes a recently introduced issue in gmt_fft.c due to the introduction of an extra bit flag: We must check for bits not using equal signs.