Check that share-path files are nonzero
This can happen when users run things like

gmt makecpt -Cturbo > turbo.cpt

because the output is created as an empty file first by the system, and then makecpt reads that instead of the master file.  The simple solution is to insist that when we search for GMT files expected in the share path, while we will allow such a file to live in the current directory it must pass the extra test that it is not empty.

All tests run and the old problem is gone.
Recently updated docs have been edited again to remove the mention of the problem.