Let modules called by modules in inset know of shrinking
See this post for background. The problem was that the coast command called another module (plot) while inside the inset section but the called module was unaware that because of the -R -J and inset size setting there was a scaling of 0.73 applied.  That meant plot was off by that factor.  THe same would happen with any other module called inside the inset section that might call another module.

The solution is to remember the computed inset scale in the API so that when the secondary module is initialized it can apply the same scaling as the mother module.