Clarify that -i will convert anything
Following the comments on the forum, I have added a note to -i to clarify that -i will override how many columns we will assume is present and barge ahead and try to convert what it finds.  This is likely to be NaN but if someone has a columne with a mix of flags containing numbers then those will presumably be concerted to numbers.  Messy and not recomme3nded, but there it is.