Add returns when gmt_M_memory fails
We have lots of calls to gmt_M_memory.  When these fails we get a detailed message of where and what the amount of memory requested was.  However, only the calling function can take further action, such as returning with or without some value.  This PR starts this process of returning when an error is encountered to prevent further carnage if just ignored.