Ensure column-wrap does not exceed periodicity
See this forum post for background.  Some months ago I fixed a bug in how we set the max width of searches per latitutude row and as we reach the pole the answer should be the full grid width, not the half.  However, because we use the periodicity count to add to any temporarily negative column values it is important that the upper limit on the search width (in nodes) at most equals that periodicity count.  This count is the same as the number of columns for pixel grids but is one less for gridline grids.  The problem reported on the forum resulted from using n_columns and not the periodicity wrap count (here n_columns-1) for the search width.