Fix wrong column selections in rotsmoother (#6795)
* Fix wrong column selections in rotsmoother

Per documentation, input format shall be lon, lat, time, angle, [weight], but code flipped time and angle.  It also did not allow a single time (since rotations usually correspond to an interval).  Now, if -T<single_t> is used we include all rotations in the calculation of a single rotation.

* Update rotsmoother.c

Fix index of last rotation outside group and also give message if not enough rotations found for a group.