Let grdflexure handle combination of -H -W
If the load has variable load density then we must use the density at each node to compute adjusted height before we determine if we now stick above sea level for the additional boost.
Also, the check for NaNs in the input grid gave an error hence we never even got to the section where the NaNs were replaced by zeros.  Now fixed.