Let gravprisms handle interfaces
An interface grid is a surface that may cross zero.  If so, the prisms we create should change sign across the zero threshold and for negative z we create a prism from x to 0 and for positive z we go from 0 to z.  If a second surface or constant is given then we have a layer and no density changes sign.  Documentation has ben updated to explain this as well adn a few comments were added elsewhere.  I also added a test on the density contrast (-D cannot be 0, but variable density from grid could be zero at times and then we skip making a dummy prism).