Improve subplot setting of -X -Y (#6844)
See this forum post for details.  Basically, If the very first GMT plot command has specific -X -Y settings then we use those settings as given else (if we are in modern mode) we use 40 inches.  However, subplot would always add -X1i -Y1i when there was no need, and then thsoe were interpreted as hardwired settings and the 40 inches offset got reset to (1,1) inch.

This PR tries harder by checking if -X -Y was in fact given and if there actulally is a shift due to clearance.