Fix documentation and handing of -J in histogram
See #6847 for background.  The handling of -J in histogram is special since only a Cartesian scaling or dimensions is allowed (i.e., map projections are not allowed).  Hence we need specific checks.  This PR does the following:

Updates the documentation for histogram to state that -J is optional in modern mode (it is required for classic mode).
Gives an error if no -J given in classic mode
Gives an error if an unsuitable projection was specified via -J.

Closes #6847.