meca -A+ssize passed wrong pointer to gmt_M_to_inch (#6856)
The pointer to the size was off by one position hence modifiers like +s5p would give

meca [WARNING]:  not a valid number and may not be decoded properly.

while +s0.2c worked since it is seen as .2c.  So we skipped the leading character and for 5p we only tried to parse p as a length.