Enhance gravprisms -D with +c for computing contrasts (#6894)
It occurred to me that it is convenient to allow -D to be used not only to set a constant density contrast for all prisms, but with a modifier (+c) instead compute the density contrast between the variable prism densities and the constant given in -D.  Otherwise I may need to compute two sets of prisms just to be able to display the densities and to compute gravity using a contrast with sea water.  This PR adds the +c modifier.  It basically reverts very recent behavior implemented after 6.4 so it has no release effect.